runtime: fix int64 reconstruction in semasleep
I had this right in one of my clients, but apparently not the one I submitted from.

Fixes 386 builds.

TBR=dfc
CC=golang-codereviews
https://golang.org/cl/138000045

+++ src/pkg/runtime/os_darwin.c
@@ -418,7 +418,7 @@ semasleep(void)
 	int32 r, secs, nsecs;
 	int64 ns;
 	
-	ns = g->m->scalararg[0] | g->m->scalararg[1]<<32;
+	ns = (int64)(uint32)g->m->scalararg[0] | (int64)(uint32)g->m->scalararg[1]<<32;
 	g->m->scalararg[0] = 0;
 	g->m->scalararg[1] = 0;
